+++ CHANGELOG.rst
@@ -26,7 +26,8 @@ Bugfixes
 ~~~~~~~~~~~~
 
 - The CA rotation procedure has been fixed once again
-  including force-renewal of the certificates and kubeconfig on Kubernetes nodes. (`!1936 <https://gitlab.com/yaook/k8s/-/merge_requests/1936>`_)
+  including force-renewal of the certificates and kubeconfig on Kubernetes nodes
+  and k8s-login for the orchestrator's kubeconfig. (`!1936 <https://gitlab.com/yaook/k8s/-/merge_requests/1936>`_)
 
 
 v9.1.9 (2025-07-07)